revise the MeCab compiler flag "MECAB_DEFAULT_RC" to set the configuration file as an empty file.
This compiler flag is the path of MeCab configuration file "mecabrc", which is "/usr/local/etc/mecabrc" defaultly if MeCab is installed in Linux. By means of this file, we could set dictionary path, output format, etc.

For our JMA development, as dictionary path is set at runtime by function calls, instead of by this configuration file, we would not use this configuration file currently.

So the comiler flag "MECAB_DEFAULT_RC" is set to an empty file, which is "/dev/null" in Linux and "NUL" in Windows.