ignore the UTF-8 BOM (0xEF 0xBB 0xBF) in iJMA analysis.
Many Windows programs use the UTF-8 byte-order mark (0xEF 0xBB 0xBF) to identify text as UTF-8 encoding,
if iJMA also analyzes the UTF-8 BOM as input, it would be output as a special charater, which result would not be expected by user.

So iJMA replace UTF-8 BOM to space characters in JMA_CType_UTF8::replaceSpaces(), and they would be ignored in the analysis afterward.