fix the problem of buffer overflow in analyzing UTF-8 text when the encoding is set as EUC-JP.
When the encoding type set by Knowledge::setEncodeType() is different with that of input text,
the problem of buffer overflow appears in JMA_CType_EUCJP::replaceSpaces() as it checks string ending by using NULL character.

To fix the problem, use the string length instead of NULL character to check its ending.